pyproject.toml
[tool.poetry]
name = "py-selfbot-starter"
version = "0.1.0"
description = ""
authors = ["0xfinder <0xfinder@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py-self" = { git = "https://github.com/dolfies/discord.py-self.git", branch = "rebase" }
requests = "2.26.0"
python-dotenv = "0.19.2"
black = "^22.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
